[event-hubs] Removed checkpointstore sample and reference it instead.
