+++ sdk/eventhub/event-hubs/samples/javascript/sample.env
@@ -4,9 +4,6 @@ EVENTHUB_NAME=<name of a single eventhub>
 EVENTHUB_FQDN=<your-eventhubs-namespace>.servicebus.windows.net
 CONSUMER_GROUP_NAME=<name of a consumer group>
 
-# Used in the receiveEventsUsingCheckpointStore.ts sample
-STORAGE_CONTAINER_NAME=<storage account connection string for checkpoints>
-
 # Used in the useWithIotHub.ts sample
 IOTHUB_EH_COMPATIBLE_CONNECTION_STRING=<connection string with EntityPath>
